requirements.txt
amqp==5.0.2
asgiref==3.2.10
autopep8==1.5.4
billiard==3.6.3.0
celery==5.0.2
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-repl==0.1.6
defusedxml==0.6.0
Django==3.1.1
django-allauth==0.42.0
django-crispy-forms==1.9.2
django-formtools==2.2
django-tinymce==3.1.0
gunicorn==20.0.4
idna==2.10
importlib-metadata==2.0.0
kombu==5.0.2
oauthlib==3.1.0
Pillow==7.2.0
prompt-toolkit==3.0.8
psycopg2==2.8.6
psycopg2-binary==2.8.6
pycodestyle==2.6.0
python-memcached==1.59
python3-openid==3.2.0
pytz==2020.1
redis==3.5.3
requests==2.24.0
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
urllib3==1.25.10
vine==5.0.0
wcwidth==0.2.5
zipp==3.4.0
